separeate form editor wrapper
