+++ core-ui/src/shared/ResourceForm/fields/Editor.scss
@@ -1,0 +1,16 @@
+.resource-form__editor {
+  .monaco-editor {
+    min-height: 200px;
+  }
+  .resource-form__editor__error {
+    inset: 0;
+
+    & > * {
+      bottom: 36px;
+      width: 50%;
+      right: 10px;
+      position: absolute;
+      opacity: 0.8;
+    }
+  }
+}
