+++ core-ui/src/shared/components/MonacoEditorESM/Editor.js
@@ -7,71 +7,8 @@ import { useTranslation } from 'react-i18next';
 import { useAutocompleteWorker } from './useAutocompleteWorker';
 import { Spinner } from 'shared/components/Spinner/Spinner';
 import { isEqual } from 'lodash';
-import * as jp from 'jsonpath';
 
 import './Editor.scss';
-
-export function EditorWrapper({
-  value,
-  onChange,
-  setValue,
-  language = 'yaml',
-  customSchemaId,
-  ...props
-}) {
-  const { t } = useTranslation();
-  const [error, setError] = useState('');
-
-  const parsedValue = React.useMemo(() => {
-    if (language === 'yaml') {
-      return jsyaml.dump(value, { noRefs: true });
-    } else if (language === 'json') {
-      return JSON.stringify(value, null, 2);
-    } else {
-      return value;
-    }
-  }, [value, language]);
-
-  // TODO, schema is lost if user deletes all the resource with the exception of one line goes to simple and returns to editor
-  const resourceSchemaId = useRef(
-    `${jp.value(value, `$.apiVersion`)}/${jp.value(value, `$.kind`)}`,
-  );
-
-  const handleChange = useCallback(
-    text => {
-      try {
-        let parsed = {};
-        if (language === 'yaml') {
-          parsed = jsyaml.load(text);
-        } else if (language === 'json') {
-          parsed = JSON.parse(text);
-        }
-        if (typeof parsed !== 'object' || !parsed) {
-          setError(t('common.create-form.object-required'));
-          return;
-        }
-        if (typeof onChange === 'function') onChange(parsed);
-        if (typeof setValue === 'function') setValue(parsed);
-
-        setError(null);
-      } catch ({ message }) {
-        // get the message until the newline
-        setError(message.substr(0, message.indexOf('\n')));
-      }
-    },
-    [onChange, setError, t, language, setValue],
-  );
-  return (
-    <Editor
-      {...props}
-      language={language}
-      value={parsedValue}
-      onChange={handleChange}
-      error={error}
-      customSchemaId={customSchemaId || resourceSchemaId.current}
-    />
-  );
-}
 
 export function Editor({
   value, // Pass value to initialize the Editor. Altering value later will not be reflected (as in an uncontrolled component).
@@ -184,18 +121,6 @@ export function Editor({
     readOnly,
     onMount,
   ]);
-
-  // useEffect(() => {
-  //   // update monaco editor value when parent modifies value
-  //   if (
-  //     JSON.stringify(value) &&
-  //     editorRef.current?.getValue() &&
-  //     JSON.stringify(value) !==
-  //       JSON.stringify(JSON.parse(editorRef.current?.getValue() || '{}'))
-  //   ) {
-  //     editorRef.current?.getModel().setValue(JSON.stringify(value));
-  //   }
-  // }, [value]);
 
   useEffect(() => {
     // refresh model on editor focus. Needed for cases when multiple editors are open simultaneously
