+++ core-ui/src/shared/components/MonacoEditorESM/Editor.scss
@@ -24,21 +24,6 @@
   overflow: hidden;
   flex-grow: 1;
   border: 1px solid var(--sapNeutralBackground);
-
-  //.monaco-editor {
-  //  min-height: 200px;
-  //}
-  //.resource-form__editor__error {
-  //  inset: 0;
-  //
-  //  & > * {
-  //    bottom: 36px;
-  //    width: 50%;
-  //    right: 10px;
-  //    position: absolute;
-  //    opacity: 0.8;
-  //  }
-  //}
 }
 
 .resource-form__legend {
